requirements.txt
torch==1.13.0
pytorch-cuda==11.6
torchvision==0.2.1
numpy==1.14.5
Pillow==6.2.1
attrdict==2.0.0
numpy==1.21.6